+++ src/model/commands/TanCommand.java
@@ -10,7 +10,7 @@ public class TanCommand extends MathCommand {
 	}
 
 	@Override
-	public double execute(int[] parameters, Turtle myTurtle, Controller view) {
+	public double execute(double[] parameters, Turtle myTurtle, Controller view) {
 		return Math.tan(parameters[0]);
 	}
 }
